Backend: Add common shared conftest and factories
- Add shared conftest and factories and use them in the shared
packages.
- Refactor with black & isort.

Refs -
