+++ backend/shared/setup.py
@@ -1,7 +1,6 @@
 import os
 
 from setuptools import find_packages, setup
-
 from shared import __version__
 
 #
