+++ backend/shared/shared/azure_adfs/tests/conftest.py
@@ -1,14 +1,1 @@
-import pytest
-from django.test import Client
-from shared.azure_adfs.tests.factories import UserFactory
-
-
-@pytest.fixture
-def user():
-    return UserFactory()
-
-
-@pytest.fixture
-def client():
-    client = Client()
-    return client
+from shared.common.tests.conftest import *  # noqa
