+++ backend/shared/shared/common/tests/factories.py
@@ -1,14 +1,12 @@
-import uuid
-
 import factory
 from django.contrib.auth import get_user_model
 
 
 class UserFactory(factory.django.DjangoModelFactory):
-    class Meta:
-        model = get_user_model()
-
-    username = factory.Sequence(lambda n: "user_%s" % uuid.uuid4())
-    email = factory.Faker("email")
     first_name = factory.Faker("first_name")
     last_name = factory.Faker("last_name")
+    username = factory.Sequence(lambda n: f"{factory.Faker('user_name')} {n}")
+    email = factory.Faker("email")
+
+    class Meta:
+        model = get_user_model()
