+++ backend/shared/shared/oidc/admin.py
@@ -1,5 +1,4 @@
 from django.contrib import admin
-
 from shared.oidc.models import EAuthorizationProfile, OIDCProfile
