+++ backend/shared/shared/oidc/jobs/hourly/clear_user_sessions.py
@@ -1,6 +1,5 @@
 from django.utils import timezone
 from django_extensions.management.jobs import HourlyJob
-
 from shared.oidc.models import EAuthorizationProfile, OIDCProfile
 from shared.oidc.services import clear_eauthorization_profiles, clear_oidc_profiles
