+++ backend/shared/shared/oidc/models.py
@@ -2,7 +2,6 @@ from django.conf import settings
 from django.db import models
 from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
-
 from shared.models.abstract_models import TimeStampedModel, UUIDModel
