+++ backend/shared/shared/oidc/tests/factories.py
@@ -1,18 +1,7 @@
 import factory
-from django.contrib.auth import get_user_model
 from django.utils import timezone
-
 from shared.oidc.models import EAuthorizationProfile, OIDCProfile
-
-
-class UserFactory(factory.django.DjangoModelFactory):
-    first_name = factory.Faker("first_name")
-    last_name = factory.Faker("last_name")
-    username = factory.Sequence(lambda n: f"{factory.Faker('user_name')} {n}")
-    email = factory.Faker("email")
-
-    class Meta:
-        model = get_user_model()
+from shared.common.tests.factories import UserFactory
 
 
 class AbstractProfileFactory(factory.django.DjangoModelFactory):
