+++ backend/shared/shared/oidc/tests/test_eauth_views.py
@@ -7,7 +7,6 @@ from django.test import override_settings
 from django.urls import reverse
 from django.utils import timezone
 from freezegun import freeze_time
-
 from shared.oidc.tests.factories import OIDCProfileFactory
 from shared.oidc.utils import get_checksum_header, get_organization_roles
