+++ backend/shared/shared/oidc/tests/test_oidc_views.py
@@ -5,7 +5,6 @@ import pytest
 from django.conf import settings
 from django.test import override_settings
 from django.urls import reverse
-
 from shared.oidc.tests.factories import OIDCProfileFactory
