+++ backend/shared/shared/oidc/utils.py
@@ -6,7 +6,6 @@ from uuid import uuid4
 import requests
 from django.conf import settings
 from django.utils import timezone
-
 from shared.oidc.auth import HelsinkiOIDCAuthenticationBackend
 from shared.oidc.models import EAuthorizationProfile, OIDCProfile
