+++ backend/shared/shared/oidc/views/hki_views.py
@@ -9,7 +9,6 @@ from django.urls import reverse
 from django.views.generic import View
 from mozilla_django_oidc.views import OIDCAuthenticationCallbackView, OIDCLogoutView
 from requests.exceptions import HTTPError
-
 from shared.oidc.models import OIDCProfile
 from shared.oidc.services import clear_oidc_profiles
 from shared.oidc.utils import get_userinfo, refresh_hki_tokens
