+++ backend/shared/shared/oidc/views/mock_views.py
@@ -4,7 +4,6 @@ from django.http import HttpResponse, HttpResponseRedirect, JsonResponse
 from django.utils.decorators import method_decorator
 from django.views.decorators.csrf import ensure_csrf_cookie
 from django.views.generic import View
-
 from shared.oidc.tests.factories import UserFactory
